Fix Error Prone warnings in `log4j-api`
This suppresses or fixes most Error Prone warning in `log4j-api`.

Only two dozens warnings are left, mainly in the
[TypeParameterUnusedInFormals](https://errorprone.info/bugpattern/TypeParameterUnusedInFormals),
which can not be removed for source compatibility reasons.

Closes #1850.

+++ log4j-api/src/main/java/org/apache/logging/log4j/message/MapMessageJsonFormatter.java
@@ -267,7 +267,7 @@ enum MapMessageJsonFormatter {;
         } else {
             final long longNumber = number.longValue();
             final double doubleValue = number.doubleValue();
-            if (Double.compare(longNumber, doubleValue) == 0) {
+            if (Double.compare((double) longNumber, doubleValue) == 0) {
                 sb.append(longNumber);
             } else {
                 sb.append(doubleValue);
